refactor(client): move route definitions
